move lazy dataset to the dataset module
